% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EPPlab.ensemble.R
\name{EPPlab.ensemble}
\alias{EPPlab.ensemble}
\title{Combining Different EPPlab results}
\usage{
EPPlab.ensemble(x, type, ...)
}
\arguments{
\item{x}{A numeric data frame or data matrix with at least two rows.}

\item{type}{The type of ensemble, see details.}

\item{...}{Tuning paramters, see \code{\link{REPPlab:EPPlab}}.}
}
\value{
The whitened data.
}
\description{
The function automatically combines the output of several \code{EPPlab} runs.
}
\details{
Instead of relying on the outcome of a single EPP-lab run one can use also this
function to run EPPlab with different indices and combine them then using the
\code{LDRTools::B2B} function.
The different predefined ensembles are defined via the \code{type} option and
possible values are: \code{clust}, \code{fast}, \code{all}, \code{friedKurt}.

All tuning parameters are as described in \code{\link{REPPlab:EPPlab}}
}
\examples{
# more observations than variables
X <- matrix(rnorm(200),ncol=4)
}
\author{
Daniel Fischer
}
\seealso{
\code{\link{LDRTools::B2B}}, \code{\link{LDRTools:AOP}}
}
\keyword{multivariate}

