% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EPPlabAgg.R
\name{EPPlabAgg}
\alias{EPPlabAgg}
\title{Function to Combine epplab Objects}
\usage{
EPPlabAgg(x, method = "cumulative", percentage = 0.95)
}
\arguments{
\item{x}{An object of class \code{epplab}.}

\item{method}{The type of method, see details}

\item{percentage}{Threshold of explained variance}
}
\value{
A list with class 'epplabagg' containing the following
components:
\item{P}{something}
\item{O}{something}
\item{k}{something}
\item{eigenvalues}{something}
}
\description{
Function that automatically determines directions of interest. Several epplab objects
can be combined
}
\details{
The parameter \code{x} can either be a single object or a list of epplab objects

Options for method are inverse, sqinverse and cum.
}
\examples{
library(tourr)
 data(olive)
 olivePP.kurt.max <-
   EPPlab(olive[,3:10],PPalg="PSO",PPindex="KurtosisMax",n.simu=10, maxiter=20)

 olivePP.fried <-
   EPPlab(olive[,3:10],PPalg="PSO",PPindex="Friedman",n.simu=10, maxiter=20)

 olivePPs <- list(olivePP.kurt.max, olivePP.fried)

 EPPlabAgg(olivePP.kurt.max)$k
 EPPlabAgg(olivePPs, "cum", 0.99)$k

 pairs(olivePP.kurt.max$x \%*\% EPPlabAgg(olivePPs, "cum", 0.99)$O,
       col=olive[,2], pch=olive[,1])


 olivAOP.sq <- EPPlabAgg(olivePPs, "inv")
 oliveProj <- olivePP.kurt.max$x \%*\% olivAOP.sq$O
 plot(density(oliveProj))
 rug(oliveProj[olive$region==1],col=1)
 rug(oliveProj[olive$region==2],col=2)
 rug(oliveProj[olive$region==3],col=3)
}
\author{
Anne Ruiz-Gazen, Daniel Fischer, Klaus Nordhausen
}
\references{
\cite{Cite What?}
}
\seealso{
\code{\link{EPPlab}}
}
\keyword{multivariate}

