% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REPPlab-package.R
\docType{data}
\name{ReliabilityData}
\alias{ReliabilityData}
\title{Reliability Data from a Manufacturer of Electronic Components}
\format{A data frame with 520 observations on 55 variables.}
\source{
The data was anonymized to keep confidentiality.
}
\description{
The data set contains 55 variables measured during the production process of
520 components. The problem for this data is to identify the produced items
with a fault.
}
\examples{
data(ReliabilityData)
summary(ReliabilityData)
}
\keyword{datasets}

