\name{REPPlab-package}
\alias{R/EPP-lab-package}
\docType{package}

\title{
  R Interface to the Java Programm EPP-lab v0.1.
}

\description{
  REPPlab is an R interface that gives access to the Java program EPP-lab which implements
  several biologically inspired optimisation algorithms and several indices for
  Exploratory Projection Pursuit (PP). The objective of optimizing PP indices
  and projecting the data on the associated one-dimensional directions is to
  detect hidden structures such as clusters or outliers in (possibly high
  dimensional) data sets.
}

\details{
  \tabular{ll}{
    Package: \tab REPPlab\cr
    Type: \tab Package\cr
    Version: \tab 0.1\cr
    Date: \tab 2012-10-08\cr
    License: \tab GPL\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Daniel Fischer, Alain Berro, Klaus Nordhausen, Anne Ruiz-Gazen

  Maintainer: Daniel Fischer <daniel.fischer@uta.fi>
}

\references{
  \cite{Larabi Marie-Sainte, S., (2011), Biologically inspired algorithms for exploratory projection pursuit, PhD thesis, University of Toulouse.}

  \cite{Ruiz-Gazen, A., Larabi M.-S. and Berro, A. (2010), Detecting multivariate outliers using projection 
   pursuit with particle swarm optimization, \emph{COMPSTAT2010}, pp. 89-98.}
}

\keyword{multivariate}

