% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EPPlabAgg.R
\name{EPPlabAgg}
\alias{EPPlabAgg}
\title{Function to Combine epplab Objects}
\usage{
EPPlabAgg(x, method = "cumulative", percentage = 0.95)
}
\arguments{
\item{x}{An object of class \code{epplab}.}

\item{method}{The type of method, see details}

\item{percentage}{Threshold of explained variance}
}
\value{
A list with class 'epplabagg' containing the following
components:
\item{P}{something}
\item{O}{something}
\item{k}{something}
\item{eigenvalues}{something}
}
\description{
Function that automatically determines directions of interest. Several epplab objects
can be combined
}
\details{
The parameter \code{x} can either be a single object or a list of epplab objects

Options for method are inverse, sqinverse and cum.
}
\examples{
# creating data with 3 outliers
n <-300
p <- 10
X <- matrix(rnorm(n*p),ncol=p)
X[1,1] <- 9
X[2,4] <- 7
X[3,6] <- 8
}
\author{
Anne Ruiz-Gazen, Daniel Fischer, Klaus Nordhausen
}
\references{
\cite{Cite What?}
}
\seealso{
\code{\link{EPPlab}}
}
\keyword{multivariate}

