% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epplab.R
\docType{methods}
\name{print.epplab}
\alias{print,epplab-method}
\alias{print-method}
\alias{print.epplab}
\title{Print an epplab Object}
\usage{
\method{print}{epplab}(x, ...)
}
\arguments{
\item{x}{Object of class \code{epplab}.}

\item{...}{Additional parameters}
}
\description{
Prints an \code{epplab} object.
}
\details{
The print function displays the result with the best value in the objective
criterion.
}
\examples{

library(tourr)
data(olive)
res <- EPPlab(olive[,3:10],PPalg="PSO",PPindex="KurtosisMin",n.simu=10, maxiter=20)
print(res)

}
\author{
Daniel Fischer
}
\keyword{methods}
\keyword{print}

